+++ src/LuaFunction.hpp
@@ -28,6 +28,7 @@ using namespace godot;
 
 namespace luagdextension {
 
+class LuaDebug;
 class VariantArguments;
 
 class LuaFunction : public LuaObjectSubclass<sol::protected_function> {
@@ -41,9 +42,11 @@ public:
 	Variant invokev(const Array& args);
 	Variant invoke(const Variant **args, GDExtensionInt arg_count, GDExtensionCallError &error);
 
+	static Variant invoke_lua(Ref<LuaFunction> f, const VariantArguments& args, bool return_lua_error);
 	static Variant invoke_lua(const sol::protected_function& f, const VariantArguments& args, bool return_lua_error);
 
 	Callable to_callable() const;
+	Ref<LuaDebug> get_debug_info() const;
 
 	const sol::protected_function& get_function() const;