+++ src/script-language/LuaScriptProperty.cpp
@@ -24,7 +24,6 @@
 #include "LuaScriptInstance.hpp"
 
 #include "../LuaCoroutine.hpp"
-#include "../LuaFunction.hpp"
 #include "../utils/VariantArguments.hpp"
 #include "../utils/VariantType.hpp"
 #include "../utils/convert_godot_lua.hpp"
@@ -70,14 +69,14 @@ static LuaScriptProperty lua_property(sol::stack_object value) {
 			property.usage &= ~PROPERTY_USAGE_STORAGE;
 		}
 		else if (auto getter = table->get<sol::optional<sol::protected_function>>("get")) {
-			property.getter = getter;
+			property.getter = LuaObject::wrap_object<LuaFunction>(*getter);
 			property.usage &= ~PROPERTY_USAGE_STORAGE;
 		}
 		if (auto setter_name = table->get<sol::optional<StringName>>("set")) {
 			property.setter_name = *setter_name;
 		}
 		else if (auto setter = table->get<sol::optional<sol::protected_function>>("set")) {
-			property.setter = setter;
+			property.setter = LuaObject::wrap_object<LuaFunction>(*setter);
 		}
 	}
 	else if (auto type = value.as<sol::optional<VariantType>>()) {
@@ -108,8 +107,8 @@ LuaScriptProperty::LuaScriptProperty(const Variant& value, const StringName& nam
 }
 
 bool LuaScriptProperty::get_value(LuaScriptInstance *self, Variant& r_value) const {
-	if (getter) {
-		r_value = LuaFunction::invoke_lua(*getter, VariantArguments(self->owner, nullptr, 0), false);
+	if (getter.is_valid()) {
+		r_value = LuaFunction::invoke_lua(getter, VariantArguments(self->owner, nullptr, 0), false);
 		return true;
 	}
 	if (!getter_name.is_empty()) {
@@ -131,8 +130,8 @@ Variant LuaScriptProperty::instantiate_default_value() const {
 }
 
 bool LuaScriptProperty::set_value(LuaScriptInstance *self, const Variant& value) const {
-	if (setter) {
-		LuaCoroutine::invoke_lua(*setter, Array::make(self->owner, value), false);
+	if (setter.is_valid()) {
+		LuaCoroutine::invoke_lua(setter, Array::make(self->owner, value), false);
 		return true;
 	}
 	else if (!setter_name.is_empty()) {