+++ src/utils/MethodBindByName.cpp
@@ -21,9 +21,12 @@
  */
 #include "MethodBindByName.hpp"
 
+#include "Class.hpp"
+#include "VariantArguments.hpp"
 #include "VariantType.hpp"
 #include "convert_godot_lua.hpp"
 
+#include <godot_cpp/classes/class_db_singleton.hpp>
 #include <godot_cpp/variant/utility_functions.hpp>
 
 namespace luagdextension {
@@ -39,9 +42,14 @@ sol::object MethodBindByName::call(sol::this_state state, const sol::stack_objec
 		luaL_error(state, "To call methods, use ':' instead of '.': `obj:%s(...)`", String(method_name).ascii().get_data());
 	}
 
-	if (self.is<VariantType>()) {
-		VariantType var_type = self.as<VariantType>();
-		return variant_static_call_string_name(state, var_type.get_type(), method_name, args);
+	if (auto var_type = self.as<sol::optional<VariantType>>()) {
+		return variant_static_call_string_name(state, var_type->get_type(), method_name, args);
+	}
+	if (auto cls = self.as<sol::optional<Class>>()) {
+		Array var_args = VariantArguments(args).get_array();
+		var_args.push_front(method_name);
+		var_args.push_front(cls->get_name());
+		return to_lua(state, ClassDBSingleton::get_singleton()->callv("class_call_static", var_args));
 	}
 	else {
 		Variant variant = to_variant(self);