+++ test/lua_tests/class_static_methods.lua
@@ -1,0 +1,2 @@
+assert(DirAccess.get_open_error, "Could not access DirAccess.get_open_error static method")
+assert(DirAccess:get_open_error(), "Could not call DirAccess.get_open_error static method")